pyproject.toml
[tool.poetry]
name = "vincent"
version = "0.1.0"
description = ""
authors = ["Joe Banks <joe@urn1350.net>"]
readme = "README.md"
packages = [{include = "vincent"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.2.2"
pynacl = "^1.5.0"
pydantic = {extras = ["dotenv"], version = "^1.10.7"}
loguru = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
